move all verification code to verifier package
